API Error Handling and Debugger
Added some better error handling for the API, and added a debug logger,
that is exported at debug.txt in the /bin/ directory after you close the
game. The debug log can be used to find errors in engine, or mod code in
the future. Although, there is more work to be done, of course.